changement de kernel vers exact predicates exact constructions pour eviter les self intersections apres les op booleennes

+++ screw.h
@@ -12,7 +12,7 @@ public:
     std::vector<Bool_op> op_bool;
 
     Mesh_CGAL mesh;
-    QVector<GLfloat> gl_data;
+    QVector<GLdouble> gl_data;
     int nb_vertices_gl_faces = 0;
     int nb_vertices_gl_lines = 0;
