+++ shape3d.h
@@ -16,7 +16,7 @@ public:
     std::vector<vertex_descriptor> v_handles;
     Mesh_CGAL mesh;
 
-    QVector<GLfloat> gl_data;
+    QVector<GLdouble> gl_data;
     int nb_vertices_gl_faces = 0;
     int nb_vertices_gl_lines = 0;
